Logs Extend the TechTrends application to log the following events:
An existing article is retrieved. The title of the article should be recorded in the log line.
A non-existing article is accessed and a 404 page is returned.
The "About Us" page is retrieved.
A new article is created. The title of the new article should be recorded in the logline.
Every log line should include the timestamp and be outputted to the STDOUT. Also, capture any Python logs at the DEBUG level.